scraping_pipeline/requirements.txt
beautifulsoup4
selenium
fastapi
lxml
uvicorn
selenium
beautifulsoup4
pymongo
webdriver-manager
lxml
python-dotenv
fastapi
uvicorn
requests
python-multipart